Move password reset logic from student to user_authn.
