+++ common/djangoapps/student/urls.py
@@ -6,7 +6,6 @@ from __future__ import absolute_import
 
 from django.conf import settings
 from django.conf.urls import url
-from django.contrib.auth.views import password_reset_complete
 
 from . import views
 
@@ -22,15 +21,6 @@ urlpatterns = [
     url(r'^change_setting$', views.change_setting, name='change_setting'),
     url(r'^change_email_settings$', views.change_email_settings, name='change_email_settings'),
 
-    # password reset in views (see below for password reset django views)
-    url(r'^account/password$', views.password_change_request_handler, name='password_change_request'),
-    url(r'^password_reset/$', views.password_reset, name='password_reset'),
-    url(
-        r'^password_reset_confirm/(?P<uidb36>[0-9A-Za-z]+)-(?P<token>.+)/$',
-        views.password_reset_confirm_wrapper,
-        name='password_reset_confirm',
-    ),
-
     url(r'^course_run/{}/refund_status$'.format(settings.COURSE_ID_PATTERN),
         views.course_run_refund_status,
         name="course_run_refund_status"),
@@ -41,14 +31,4 @@ urlpatterns = [
         name='activate_secondary_email'
     ),
 
-]
-
-# password reset django views (see above for password reset views)
-urlpatterns += [
-    # TODO: Replace with Mako-ized views
-    url(
-        r'^password_reset_complete/$',
-        password_reset_complete,
-        name='password_reset_complete',
-    ),
 ]
